% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultEvidenceSynthesisConfig}
\alias{createDefaultEvidenceSynthesisConfig}
\title{createDefaultEvidenceSynthesisConfig}
\usage{
createDefaultEvidenceSynthesisConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the Evidence Synthesis config specification (depreciated)
}
\details{
User specifies the settings to create a default config for a Evidence Synthesis module
}
\seealso{
Other ModuleConfig:
\code{\link[=createDefaultAboutConfig]{createDefaultAboutConfig()}},
\code{\link[=createDefaultCharacterizationConfig]{createDefaultCharacterizationConfig()}},
\code{\link[=createDefaultCohortDiagnosticsConfig]{createDefaultCohortDiagnosticsConfig()}},
\code{\link[=createDefaultCohortGeneratorConfig]{createDefaultCohortGeneratorConfig()}},
\code{\link[=createDefaultCohortMethodConfig]{createDefaultCohortMethodConfig()}},
\code{\link[=createDefaultDatasourcesConfig]{createDefaultDatasourcesConfig()}},
\code{\link[=createDefaultEstimationConfig]{createDefaultEstimationConfig()}},
\code{\link[=createDefaultHomeConfig]{createDefaultHomeConfig()}},
\code{\link[=createDefaultPhevaluatorConfig]{createDefaultPhevaluatorConfig()}},
\code{\link[=createDefaultPredictionConfig]{createDefaultPredictionConfig()}},
\code{\link[=createDefaultReportConfig]{createDefaultReportConfig()}},
\code{\link[=createDefaultSccsConfig]{createDefaultSccsConfig()}},
\code{\link[=createModuleConfig]{createModuleConfig()}}
}
\concept{ModuleConfig}
