% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultSccsConfig}
\alias{createDefaultSccsConfig}
\title{createDefaultSccsConfig}
\usage{
createDefaultSccsConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the SCCS config specification (depreciated)
}
\details{
User specifies the settings to create a default config for a SCCS module
}
\seealso{
Other ModuleConfig:
\code{\link[=createDefaultAboutConfig]{createDefaultAboutConfig()}},
\code{\link[=createDefaultCharacterizationConfig]{createDefaultCharacterizationConfig()}},
\code{\link[=createDefaultCohortDiagnosticsConfig]{createDefaultCohortDiagnosticsConfig()}},
\code{\link[=createDefaultCohortGeneratorConfig]{createDefaultCohortGeneratorConfig()}},
\code{\link[=createDefaultCohortMethodConfig]{createDefaultCohortMethodConfig()}},
\code{\link[=createDefaultDatasourcesConfig]{createDefaultDatasourcesConfig()}},
\code{\link[=createDefaultEstimationConfig]{createDefaultEstimationConfig()}},
\code{\link[=createDefaultEvidenceSynthesisConfig]{createDefaultEvidenceSynthesisConfig()}},
\code{\link[=createDefaultHomeConfig]{createDefaultHomeConfig()}},
\code{\link[=createDefaultPhevaluatorConfig]{createDefaultPhevaluatorConfig()}},
\code{\link[=createDefaultPredictionConfig]{createDefaultPredictionConfig()}},
\code{\link[=createDefaultReportConfig]{createDefaultReportConfig()}},
\code{\link[=createModuleConfig]{createModuleConfig()}}
}
\concept{ModuleConfig}
