% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadConfig.R
\name{loadConfig}
\alias{loadConfig}
\title{loadConfig}
\usage{
loadConfig(configLocation)
}
\arguments{
\item{configLocation}{The location of the config json file}
}
\value{
An R list with the config settings
}
\description{
Load the config specification
}
\details{
User specifies the config file and this function returns the R list with the config
}
\seealso{
Other Config:
\code{\link[=addModuleConfig]{addModuleConfig()}},
\code{\link[=initializeModuleConfig]{initializeModuleConfig()}},
\code{\link[=saveConfig]{saveConfig()}}
}
\concept{Config}
