% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiShinyAppBuilder.R
\docType{package}
\name{OhdsiShinyAppBuilder}
\alias{OhdsiShinyAppBuilder-package}
\alias{OhdsiShinyAppBuilder}
\title{OhdsiShinyAppBuilder}
\description{
A package for creating viewers for the OHDSI shiny apps
}
\author{
\strong{Maintainer}: Jenna Reps \email{jreps@its.jnj.com}

Authors:
\itemize{
  \item Jenna Reps \email{jreps@its.jnj.com}
  \item Nathan Hall
  \item Josh Ide
  \item Jamie Gibert
}

}
