% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLoadConfig.R
\name{saveConfig}
\alias{saveConfig}
\title{saveConfig}
\usage{
saveConfig(configList, configLocation)
}
\arguments{
\item{configList}{An R list with the config settings}

\item{configLocation}{The location to save the config json file}
}
\value{
An R list with the config settings
}
\description{
Save the R list with the config specification as a json
}
\details{
User saves the R list with the config as a json file
}
\seealso{
Other Config:
\code{\link[=addModuleConfig]{addModuleConfig()}},
\code{\link[=initializeModuleConfig]{initializeModuleConfig()}},
\code{\link[=loadConfig]{loadConfig()}}
}
\concept{Config}
