% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateConfig.R
\name{createDefaultCohortGeneratorConfig}
\alias{createDefaultCohortGeneratorConfig}
\title{createDefaultCohortGeneratorConfig}
\usage{
createDefaultCohortGeneratorConfig()
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the cohort generator config specification
}
\details{
User specifies the settings to create a default config for a cohort generator module
}
\seealso{
Other ModuleConfig: 
\code{\link{createDefaultAboutConfig}()},
\code{\link{createDefaultCharacterizationConfig}()},
\code{\link{createDefaultCohortDiagnosticsConfig}()},
\code{\link{createDefaultCohortMethodConfig}()},
\code{\link{createDefaultDatasourcesConfig}()},
\code{\link{createDefaultEstimationConfig}()},
\code{\link{createDefaultEvidenceSynthesisConfig}()},
\code{\link{createDefaultHomeConfig}()},
\code{\link{createDefaultPhevaluatorConfig}()},
\code{\link{createDefaultPredictionConfig}()},
\code{\link{createDefaultReportConfig}()},
\code{\link{createDefaultSccsConfig}()},
\code{\link{createModuleConfig}()}
}
\concept{ModuleConfig}
