% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultDatabaseSettings.R
\name{createDefaultResultDatabaseSettings}
\alias{createDefaultResultDatabaseSettings}
\title{createDefaultResultDatabaseSettings}
\usage{
createDefaultResultDatabaseSettings(
  schema = "main",
  vocabularyDatabaseSchema = schema,
  cgTablePrefix = "cg_",
  cgTable = "cohort_definition",
  databaseTable = "DATABASE_META_DATA",
  databaseTablePrefix = "",
  cdTablePrefix = "cd_",
  cTablePrefix = "c_",
  incidenceTablePrefix = "ci_",
  plpTablePrefix = "plp_",
  cmTablePrefix = "cm_",
  sccsTablePrefix = "sccs_",
  sccTablePrefix = "scc_",
  esTablePrefix = "es_",
  pvTablePrefix = "pv_"
)
}
\arguments{
\item{schema}{The schema containing the OHDSI study results}

\item{vocabularyDatabaseSchema}{The schema containing the vocabulary database}

\item{cgTablePrefix}{The table prefix for cohort generator tables}

\item{cgTable}{The table with the cohort definitions in cohort generator}

\item{databaseTable}{The table with the database details}

\item{databaseTablePrefix}{The prefix for the databaseTable}

\item{cdTablePrefix}{The table prefix for the cohort diagnostic results}

\item{cTablePrefix}{The table prefix for the characterization results}

\item{incidenceTablePrefix}{The table prefix for the incidence results}

\item{plpTablePrefix}{The table prefix for the prediction results}

\item{cmTablePrefix}{The table prefix for the cohort method results}

\item{sccsTablePrefix}{The table prefix for the self controlled case series results}

\item{sccTablePrefix}{The table prefix for the self controlled cohort results}

\item{esTablePrefix}{The table prefix for the evidence synthesis method results}

\item{pvTablePrefix}{The table prefix for the phevaluator results}
}
\value{
An R list with the module config settings
}
\description{
Create an R list with the default table prefixes and tables
}
\details{
User specifies the schema name and can override default table prefixes and table names
}
\concept{DefaultResultDatabase}
