% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiShinyAppBuilder.R
\docType{package}
\name{OhdsiShinyAppBuilder}
\alias{OhdsiShinyAppBuilder-package}
\alias{OhdsiShinyAppBuilder}
\title{OhdsiShinyAppBuilder}
\description{
A package for creating viewers for the OHDSI shiny apps
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/OhdsiShinyAppBuilder}
  \item \url{https://github.com/OHDSI/OhdsiShinyAppBuilder}
  \item Report bugs at \url{https://github.com/OHDSI/OhdsiShinyAppBuilder/issues}
}

}
\author{
\strong{Maintainer}: Jenna Reps \email{jreps@its.jnj.com}

Authors:
\itemize{
  \item Nathan Hall
  \item Josh Ide
  \item Jamie Gibert
}

}
